+++ test/core/common/CMakeLists.txt
@@ -23,7 +23,3 @@ addtest(blob_test
 target_link_libraries(blob_test
   blob
   )
-
-addtest(event_emitter_test
-    event_emitter_test.cpp
-    )
