+++ test/core/libp2p/CMakeLists.txt
@@ -9,3 +9,4 @@ add_subdirectory(peer)
 add_subdirectory(transport)
 add_subdirectory(crypto)
 add_subdirectory(upgrader)
+add_subdirectory(event)
