+++ test/core/libp2p/event/CMakeLists.txt
@@ -1,0 +1,8 @@
+#
+# Copyright Soramitsu Co., Ltd. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+addtest(event_emitter_test
+    event_emitter_test.cpp
+    )
